requirements.txt
pycryptodome==3.10.1
pycrypto==3.10.1

